Reorganized and split into smaller modules.
