+++ waspc/waspc.cabal
@@ -546,9 +546,13 @@ library cli-lib
     Wasp.Cli.Command.CreateNewProject.ArgumentsParser
     Wasp.Cli.Command.CreateNewProject.Common
     Wasp.Cli.Command.CreateNewProject.ProjectDescription
-    Wasp.Cli.Command.CreateNewProject.StarterTemplates
+    Wasp.Cli.Command.CreateNewProject.StarterTemplates.Common
+    Wasp.Cli.Command.CreateNewProject.StarterTemplates.FeaturedStarterTemplates
     Wasp.Cli.Command.CreateNewProject.StarterTemplates.GhRepo
     Wasp.Cli.Command.CreateNewProject.StarterTemplates.Local
+    Wasp.Cli.Command.CreateNewProject.StarterTemplates.Skeleton
+    Wasp.Cli.Command.CreateNewProject.StarterTemplates.StarterTemplate
+    Wasp.Cli.Command.CreateNewProject.StarterTemplates.StarterTemplateId
     Wasp.Cli.Command.CreateNewProject.StarterTemplates.Templating
     Wasp.Cli.Command.Db
     Wasp.Cli.Command.Db.Migrate
